Fix: Update extensions of file to .jsx for consistency
