+++ client/src/components/About/TextWrapper.jsx
